+++ lib/aws-sdk-rails.rb
@@ -9,6 +9,7 @@ require_relative 'aws/rails/sqs_active_job/job_runner'
 
 require_relative 'action_dispatch/session/dynamodb_store'
 require_relative 'active_job/queue_adapters/amazon_sqs_adapter'
+require_relative 'active_job/queue_adapters/amazon_sqs_async_adapter'
 
 require_relative 'generators/aws_record/base'