+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 Unreleased Changes
 ------------------
 
+* Feature - Add a non-blocking async ActiveJob adapter: `:amazon_sqs_async`.  
+
+* Issue - Fix bug in default for backpressure config.
+
 3.3.0 (2020-12-01)
 ------------------
 
-* Feature - Add `aws-record` as a dependency, a rails generator for `aws-record` models, and a rake task for table migrations. 
+* Feature - Add `aws-record` as a dependency, a rails generator for `aws-record` models, and a rake task for table migrations.
 
 * Feature - Add AWS SQS ActiveJob - A lightweight, SQS backend for ActiveJob.